myproject/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-admin-interface==0.29.2
django-colorfield==0.11.0
django-jalali==7.2.0
django-jalali-date==1.1.3
django-widget-tweaks==1.5.0
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
jalali_core==1.0.0
jdatetime==5.0.0
lxml==5.3.0
melipayamakwebservice==1.0.1
multidict==6.1.0
mymelipayamak==1.0.0
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.1
python-slugify==8.0.4
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==75.6.0
sqlparse==0.5.2
text-unidecode==1.3
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.3
zeep==4.3.1
